feat: 메시지 정렬 추가

See merge request s11-final/S11P31A109!112

+++ backend/Bobi/src/main/java/com/haneolenae/bobi/domain/custom/repository/CustomTemplateRepository.java
@@ -32,8 +32,8 @@ public interface CustomTemplateRepository extends JpaRepository<CustomTemplate, 
 		"AND (:templateSearch IS NULL OR " +
 		"     t.title LIKE CONCAT('%', :templateSearch, '%') OR t.content LIKE CONCAT('%', :templateSearch, '%'))")
 	Page<CustomTemplate> findTemplates(Pageable pageable,
-		@Param("templateTags") List<Integer> templateTags,
-		@Param("templateCustomer") List<Integer> templateCustomer,
+		@Param("templateTags") List<Long> templateTags,
+		@Param("templateCustomer") List<Long> templateCustomer,
 		@Param("templateSearch") String templateSearch,
 		@Param("memberId") long memberId);
